[Torch] move DenseResource handling into torch.vtensor's canonicalize
